Fix issue with Parsson, move things around, update test
